# go-gorm-swagger-zap/dal/model/book.gen.go
// Code generated by gorm.io/gen. DO NOT EDIT.
// Code generated by gorm.io/gen. DO NOT EDIT.
// Code generated by gorm.io/gen. DO NOT EDIT.

package model

import (
	"time"
)

const TableNameBook = "book"

// Book 书籍表
type Book struct {
	ID          int64     `gorm:"column:id;primaryKey;autoIncrement:true;comment:主键" json:"id"`                   // 主键
	Title       string    `gorm:"column:title;not null;comment:书籍名称" json:"title"`                                // 书籍名称
	Author      string    `gorm:"column:author;not null;comment:作者" json:"author"`                                // 作者
	Price       int32     `gorm:"column:price;not null;comment:价格" json:"price"`                                  // 价格
	PublishDate time.Time `gorm:"column:publish_date;default:CURRENT_TIMESTAMP;comment:出版日期" json:"publish_date" swaggertype:"string" format:"date-time"` // 出版日期
	CreateAt    time.Time `gorm:"column:create_at;default:CURRENT_TIMESTAMP;comment:创建时间" json:"create_at" swaggertype:"string" format:"date-time"`       // 创建时间
	UpdateAt    time.Time `gorm:"column:update_at;default:CURRENT_TIMESTAMP;comment:更新时间" json:"update_at" swaggertype:"string" format:"date-time"`       // 更新时间
}

// TableName Book's table name
func (*Book) TableName() string {
	return TableNameBook
}