# go-gorm-swagger-zap/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Book API+SQL 增删改查",
        "title": "Book API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "龙应华",
            "url": "http://www.swagger.io/support",
            "email": "542791872@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "10.40.3.9:8080",
    "basePath": "/api/v1",
    "paths": {
        "/create/books1": {
            "post": {
                "description": "前端传入book书籍相关信息直接保存book对象信息的接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "单本书籍存储"
                ],
                "summary": "存储book书籍信息",
                "parameters": [
                    {
                        "description": "book书籍的title，author，price相关信息",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "插入数据成功提示",
                        "schema": {
                            "$ref": "#/definitions/ret.Result1"
                        }
                    },
                    "400": {
                        "description": "错误提示",
                        "schema": {
                            "$ref": "#/definitions/ret.Result1"
                        }
                    }
                }
            }
        },
        "/create/books2": {
            "post": {
                "description": "前端传入book书籍相关信息直接保存book对象信息的接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "单本书籍存储"
                ],
                "summary": "存储book书籍信息",
                "parameters": [
                    {
                        "description": "book书籍的title，author，price相关信息",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "插入数据成功提示",
                        "schema": {
                            "$ref": "#/definitions/ret.Result1"
                        }
                    },
                    "400": {
                        "description": "错误提示",
                        "schema": {
                            "$ref": "#/definitions/ret.Result1"
                        }
                    }
                }
            }
        },
        "/create/books3": {
            "post": {
                "description": "前端传入book书籍相关信息直接保存book对象信息的接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "多本书籍存储"
                ],
                "summary": "存储book书籍信息",
                "parameters": [
                    {
                        "description": "book书籍的title，author，price相关信息",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RequestPayloadBook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "插入数据成功提示",
                        "schema": {
                            "$ref": "#/definitions/ret.Result1"
                        }
                    },
                    "400": {
                        "description": "错误提示",
                        "schema": {
                            "$ref": "#/definitions/ret.Result1"
                        }
                    }
                }
            }
        },
        "/create/books4": {
            "post": {
                "description": "前端传入book书籍相关信息直接保存book对象信息的接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "多本书籍存储"
                ],
                "summary": "存储book书籍信息",
                "parameters": [
                    {
                        "description": "book书籍的title，author，price相关信息",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RequestPayloadBook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "插入数据成功提示",
                        "schema": {
                            "$ref": "#/definitions/ret.Result1"
                        }
                    },
                    "400": {
                        "description": "错误提示",
                        "schema": {
                            "$ref": "#/definitions/ret.Result1"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "作者",
                    "type": "string"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string",
                    "format": "date-time"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "price": {
                    "description": "价格",
                    "type": "integer"
                },
                "publish_date": {
                    "description": "出版日期",
                    "type": "string",
                    "format": "date-time"
                },
                "title": {
                    "description": "书籍名称",
                    "type": "string"
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "model.RequestPayloadBook": {
            "type": "object",
            "properties": {
                "books": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Book"
                    }
                }
            }
        },
...
        "ret.Result1": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}